Cambios en BD, compras y ventas

+++ vistas/scripts/bodegaTienda.js
@@ -271,33 +271,47 @@ function agregarDetalle(idarticulo,articulo,stock,stocktienda)
 {
     
     var cantidad = 1;
+    if(!yaExiste(idarticulo)){
+        if(idarticulo != "")
+        {
+            var fila = '<tr class="filas" id="fila'+cont+'"> ' +
+                        '<td>'+
+                            '<button type="button" class="btn btn-danger" onclick="eliminarDetalle('+cont+')">X</button>'+
+                        '</td>'+
+                        '<td>' +
+                            '<input type="hidden" name="idarticulo'+cont+'" id="idarticulo'+cont+'" value="'+idarticulo+'">'+
+                            articulo +
+                        '</td>'+
+                        '<td>' +
+                                '<input type="hidden" name="stock'+cont+'" id="stock'+cont+'" value="'+stock+'">'+
+                                '<input type="hidden" name="stockTienda'+cont+'" id="stockTienda'+cont+'" value="'+stocktienda+'">'+
+                            '<input type="number" name="cantidad'+cont+'" id="cantidad'+cont+'" onchange="calcularTotales()" min="1" max="'+stock+'"  value="'+cantidad+'"  >'+
+                        '</td>'
+                    '</tr>';
 
-    if(idarticulo != "")
-    {
-        var fila = '<tr class="filas" id="fila'+cont+'"> ' +
-                      '<td>'+
-                           '<button type="button" class="btn btn-danger" onclick="eliminarDetalle('+cont+')">X</button>'+
-                       '</td>'+
-                      '<td>' +
-                          '<input type="hidden" name="idarticulo'+cont+'" id="idarticulo'+cont+'" value="'+idarticulo+'">'+
-                           articulo +
-                       '</td>'+
-                      '<td>' +
-                            '<input type="hidden" name="stock'+cont+'" id="stock'+cont+'" value="'+stock+'">'+
-                            '<input type="hidden" name="stockTienda'+cont+'" id="stockTienda'+cont+'" value="'+stocktienda+'">'+
-                          '<input type="number" name="cantidad'+cont+'" id="cantidad'+cont+'" onchange="calcularTotales()" min="1" max="'+stock+'"  value="'+cantidad+'"  >'+
-                       '</td>'
-                   '</tr>';
+            cont++;
+            detalles++;
+            $("#detalles").append(fila);
+            calcularTotales(); 
+        }
+        else
+        {
+            alert("Error al ingresar el detalle, revisar los ddatos del articulo");
+        }
+    }
+    
+}
 
-        cont++;
-        detalles++;
-        $("#detalles").append(fila);
-        calcularTotales(); 
+function yaExiste(idarticulo)
+{
+    for(var i=0; i<cont;i++){
+        var id= $('#idarticulo'+i).val();
+        if(id==idarticulo){
+            bootbox.alert('Este producto ya está agregado');
+            return true;
+        }
     }
-    else
-    {
-        alert("Error al ingresar el detalle, revisar los ddatos del articulo");
-    }
+    return false;
 }
 
 function eliminarDetalle(indice)
