+++ vistas/scripts/ingreso.js
@@ -44,16 +44,21 @@ function limpiar()
 {
     $("#idproveedor").val(0);
     $("#idproveedor").selectpicker('refresh');
-    //$("#usuario").val("");
+    $("#idtienda").val(0);
+    $("#idtienda").selectpicker('refresh');
+    $("#idmoneda").val(0);
+    $("#idmoneda").selectpicker('refresh');
+    $("#idestado").val(0);
+    $("#idestado").selectpicker('refresh');
     $("#fecha_hora").val("");
     $("#total").val("");
-    $("#idcompraencabezado").val("");
+    $("#impuestos").val("");
     $("#btnAgregarArt").prop("disabled",true);
     $("#btnAgregarArt").prop("title",'Seleccione primero el proveedor y la tienda');
     $("#total_compra").val("");
     $(".filas").remove();
     $("#total").html(0);
-
+    evaluar();
     //obtenemos la fecha actual
     var now = new Date();
     var day = ("0" + now.getDate()).slice(-2);
@@ -80,12 +85,24 @@ function mostrarform(flag)
         $("#btnCancelar").show();
         detalles = 0;
         $("#btnAgregarArt").show();
+        $("#estado").css('display',"none");
+        $("#referencia").css('display',"inline");
+        $("#idcompraencabezado").css('display',"inline");
+        $("#idproveedor").prop("disabled",false);
+        $("#idtienda").prop("disabled",false);
+        $("#fecha_hora").prop("disabled",false);
+        $("#usuario").prop("disabled",false);
+        $("#usuario").prop("disabled",false);
+        $("#idmoneda").prop("disabled",false);
+        $("#impuestos").prop("disabled",false);
     }
     else
     {
         $("#listadoregistros").show();
         $("#formularioregistros").hide();
         $("#btnagregar").show();
+        $("#referencia").css('display',"none");
+        $("#idcompraencabezado").css('display',"none");
     }
 }
 
@@ -121,13 +138,15 @@ function combosSelected(){
     }
 }
 
-function desbloquear()
+function desbloquear(mensaje)
 {
     $("#idproveedor").prop("disabled",false);
     $("#idtienda").prop("disabled",false);
    //falta limpiar
    limpiar();
-   bootbox.alert("Datos limpiados\n Proveedor y tienda desbloqueados");
+   if(mensaje){
+    bootbox.alert("Datos limpiados\n Proveedor y tienda desbloqueados");
+   }
 }
 
 //Funcion listar
@@ -194,10 +213,11 @@ function listarArticulos()
 //funcion para guardar o editar
 function guardaryeditar(e)
 {
+    var fecha = $("#fecha_hora").val();
     var idproveedor = $("#idproveedor").val();
     var idtienda = $("#idtienda").val();
     var impuesto = $("#impuestos").val();
-    var usuario = $("#usuario").val();
+    var usuario = $("#idusuario").val();
     var moneda = $("#idmoneda").val();
     var total = $("#total_compra").val();
     var articulos = [];
@@ -206,51 +226,73 @@ function guardaryeditar(e)
         articulos.push($('#idarticulo'+i).val());
         cantidad.push($('#cantidad'+i).val());
     }
-    $.post(
-        "../ajax/tienda.php?op=guardaryeditar",
-        {idproveedor:idproveedor,idtienda:idtienda,articulos:articulos,cantidad:cantidad,impuesto:impuesto, usuario:usuario, moneda:moneda, total:total},
-        function(e)
-        {
-            bootbox.alert(e);
-            desbloquear(false);
-        }
-    );
-
-    //limpiar();*/
+    if (impuesto==''){
+        bootbox.alert('Debe de llenar el impuesto');
+    }else if(moneda==0 || moneda==null || moneda==''){
+        bootbox.alert('Debe de seleccionar la moneda');
+    }else{
+        $.post(
+            "../ajax/ingreso.php?op=guardaryeditar",
+            {fecha:fecha,idproveedor:idproveedor,idtienda:idtienda,articulos:articulos,cantidad:cantidad,impuesto:impuesto, usuario:usuario, moneda:moneda, total:total},
+            function(e)
+            {
+                bootbox.alert(e);
+                desbloquear(false);
+            }
+        );
+    }
+    detalles=0;
+    desbloquear();
 }
 
 function mostrar(idcompraencabezado)
 {
-    /*$.post(
+    $.post(
         "../ajax/ingreso.php?op=mostrar",{idcompraencabezado:idcompraencabezado},function(data,status)
         {
             
             data = JSON.parse(data);            
             mostrarform(true);
 
-            $("#proveedor").val(data.proveedor);
-            
+            $("#idproveedor").val(data.idproveedor);
+            $("#idproveedor").selectpicker('refresh');
+            $("#idtienda").val(data.idtienda);
+            $("#idtienda").selectpicker('refresh');
+            $("#idmoneda").val(data.idtipomoneda);
+            $("#idmoneda").selectpicker('refresh');
+            $("#idestado").val(parseInt(data.estado)+1);
+            $("#idestado").selectpicker('refresh');
+
+            $("#impuestos").val(data.impuesto);            
             $("#usuario").val(data.usuario);
             $("#fecha_hora").val(data.fecha);
-            $("#total").val('Q '+data.total);            
-            $("#idcompraencabezado").val(data.idcompraencabezado); 
+            $("#total").html('Q '+data.total);            
+            $("#idcompraencabezado").html(data.idcompraencabezado); 
             
             //Ocultar y mostrar botones
             $("#btnGuardar").hide();
             $("#btnCancelar").show();
             $("#btnAgregarArt").hide();
+            $("#estado").css('display',"block");
+            $("#idproveedor").prop("disabled",true);
+            $("#idtienda").prop("disabled",true);
+            $("#fecha_hora").prop("disabled",true);
+            $("#usuario").prop("disabled",true);
+            $("#usuario").prop("disabled",true);
+            $("#idmoneda").prop("disabled",true);
+            $("#impuestos").prop("disabled",true);
 
             $.post(
                 "../ajax/ingreso.php?op=listarDetalle&id="+idcompraencabezado,function(r)
                 {
-                    console.log(r);
+                    //console.log(r);
                     $("#detalles").html("");
                     $("#detalles").html(r);
                 }
             );
 
         }
-    );*/
+    );
 
 }
 
@@ -295,41 +337,42 @@ function marcarImpuesto()
 
 function agregarDetalle(idarticulo,articulo,preciocompra,precioventa)
 {
+    if(!yaExiste(idarticulo)){
+        if(idarticulo != "")
+        {
+            var subtotal = 1 * preciocompra;
+            var fila = '<tr class="filas" id="fila'+cont+'"> ' +
+                        '<td>'+
+                            '<button type="button" class="btn btn-danger" onclick="eliminarDetalle('+cont+')">X</button>'+
+                        '</td>'+
+                        '<td>' +
+                            '<input type="hidden" name="idarticulo'+cont+'" id="idarticulo'+cont+'" value="'+idarticulo+'">'+
+                            articulo +
+                        '</td>'+
+                        '<td>' +
+                            '<input type="number" name="cantidad'+cont+'" id="cantidad'+cont+'" onchange="modificarSubtotales()" min="1" value=1>'+
+                        '</td>'+
+                        '<td>' +
+                                '<input type="hidden" name="precio'+cont+'" id="precio'+cont+'" value="'+preciocompra+'">'+
+                                preciocompra+
+                        '</td>'+
+                        '<td>' +
+                                precioventa+
+                        '</td>'+
+                        '<td>' +
+                            '<span name="subtotal'+cont+'" id="subtotal'+cont+'">'+subtotal+'</span>'+
+                        '</td>'+
+                    '</tr>';
 
-    if(idarticulo != "")
-    {
-        var subtotal = 1 * preciocompra;
-        var fila = '<tr class="filas" id="fila'+cont+'"> ' +
-                      '<td>'+
-                           '<button type="button" class="btn btn-danger" onclick="eliminarDetalle('+cont+')">X</button>'+
-                       '</td>'+
-                      '<td>' +
-                          '<input type="hidden" name="idarticulo'+cont+'" id="idarticulo'+cont+'" value="'+idarticulo+'">'+
-                           articulo +
-                       '</td>'+
-                      '<td>' +
-                          '<input type="number" name="cantidad'+cont+'" id="cantidad'+cont+'" onchange="modificarSubtotales()" min="1" value=1>'+
-                       '</td>'+
-                      '<td>' +
-                            '<input type="hidden" name="precio'+cont+'" id="precio'+cont+'" value="'+preciocompra+'">'+
-                            preciocompra+
-                       '</td>'+
-                      '<td>' +
-                            precioventa+
-                       '</td>'+
-                      '<td>' +
-                          '<span name="subtotal'+cont+'" id="subtotal'+cont+'">'+subtotal+'</span>'+
-                       '</td>'+
-                   '</tr>';
-
-        cont++;
-        detalles++;
-        $("#detalles").append(fila);
-        modificarSubtotales(); 
-    }
-    else
-    {
-        alert("Error al ingresar el detalle, revisar los ddatos del articulo");
+            cont++;
+            detalles++;
+            $("#detalles").append(fila);
+            modificarSubtotales(); 
+        }
+        else
+        {
+            alert("Error al ingresar el detalle, revisar los ddatos del articulo");
+        }
     }
 }
 
@@ -358,6 +401,18 @@ function calcularTotales()
     $("#total_compra").val(total);
 
     evaluar();
+}
+
+function yaExiste(idarticulo)
+{
+    for(var i=0; i<cont;i++){
+        var id= $('#idarticulo'+i).val();
+        if(id==idarticulo){
+            bootbox.alert('Este producto ya está agregado');
+            return true;
+        }
+    }
+    return false;
 }
 
 function evaluar()
